this commit will move all the auth endpoints into their own file
