+++ backend/prisma/schema.prisma
@@ -17,7 +17,7 @@ model User {
   id                Int          @id @default(autoincrement()) // uid generated by firebase
   firebaseId        String       @unique
   email             String       @unique
-  intolerances     String[]
+  intolerances      String[]
   diets             String[]
   ingredientsOnHand Ingredient[]
   groceryList       GroceryList? // store the grocery list for a user
@@ -33,9 +33,8 @@ model Ingredient {
   productImage         String?
   expiration           DateTime?
   estimatedCost        Decimal?
-  userId               Int // if ingredient is an ingredient a user owns
-  user                 User         @relation(fields: [userId], references: [id])
   groceryListId        Int?
+  users                User[]
   groceryList          GroceryList? @relation(fields: [groceryListId], references: [id])
   recipeId             Int
   recipe               Recipe       @relation(fields: [recipeId], references: [id])
